+++ mod/datalynx/entries_class.php
@@ -508,7 +508,7 @@ class datalynx_entries {
      * Process entries when after editing content for saving into db
      *
      * @param string $action
-     * @param string $eids
+     * @param string||array $eids
      * @param null $data
      * @param bool $confirmed
      * @return array notificationstrings, list of processed ids
@@ -516,7 +516,7 @@ class datalynx_entries {
      * @throws dml_exception
      * @throws moodle_exception
      */
-    public function process_entries(string $action, string $eids, $data = null, bool $confirmed = false): array {
+    public function process_entries(string $action, $eids, $data = null, bool $confirmed = false): array {
         global $DB, $USER, $OUTPUT, $PAGE;
         $dl = $this->datalynx;
         $errorstring = '';
@@ -541,7 +541,7 @@ class datalynx_entries {
                 $addcount = 0;
                 $addmax = $dl->data->maxentries;
                 $perinterval = ($dl->data->intervalcount > 1);
-                if ($addmax != -1 and has_capability('mod/datalynx:manageentries', $dl->context)) {
+                if ($addmax != -1 && has_capability('mod/datalynx:manageentries', $dl->context)) {
                     $addmax = -1;
                 } else {
                     if ($addmax != -1) {
@@ -554,7 +554,7 @@ class datalynx_entries {
                     $entry = new stdClass();
 
                     // Existing entry from view.
-                    if ($eid > 0 and isset($this->_entries[$eid])) {
+                    if ($eid > 0 && isset($this->_entries[$eid])) {
                         $entries[$eid] = $this->_entries[$eid];
 
                         // TODO existing entry *not* from view (import).
@@ -596,14 +596,14 @@ class datalynx_entries {
                     ) {
                         unset($entries[$eid]);
                         $capname = get_string('datalynx:approve', 'mod_datalynx');
-                        $errorstring .= get_string('missingrequiredcapability', 'webservice', $capname) ;
+                        $errorstring .= get_string('missingrequiredcapability', 'webservice', $capname);
                         $errorstring .= get_string('affectedid', 'mod_datalynx', $eid) . '<br>';
                         // Filter managable entries.
                     } else {
                         if (!$dl->user_can_manage_entry($entry)) {
                             unset($entries[$eid]);
                             $capname = get_string('updateentry', 'mod_datalynx');
-                            $errorstring .= get_string('missingrequiredcapability', 'webservice', $capname) ;
+                            $errorstring .= get_string('missingrequiredcapability', 'webservice', $capname);
                             $errorstring .= get_string('affectedid', 'mod_datalynx', $eid) . '<br>';
                         }
                     }