+++ mod/datalynx/rule/rule_form.php
@@ -101,7 +101,7 @@ class datalynx_rule_form extends moodleform {
         // If we have selected entry updated, add a new UI when the instance includes a checkbox.
         if($checkboxes = $this->_df->get_fields_by_type('checkbox', true)) {
             $checkboxes = array('0' => get_string('noselection', 'datalynx')) + $checkboxes;
-            $mform->addElement('select', 'param5', get_string('triggerspecificevent', 'datalynxrule_eventnotification'), $checkboxes);      
+            $mform->addElement('select', 'param5', get_string('triggerspecificevent', 'datalynxrule_eventnotification'), $checkboxes);
         }
         $this->rule_definition();
 
@@ -115,10 +115,12 @@ class datalynx_rule_form extends moodleform {
      * @param array|stdClass $data
      */
     public function set_data($data) {
-        $selectedevents = unserialize($data->param1);
-        if ($selectedevents) {
-            foreach ($selectedevents as $eventname) {
-                $data->$eventname = true;
+        if (!empty($data->param1)) {
+            $selectedevents = unserialize($data->param1);
+            if ($selectedevents) {
+                foreach ($selectedevents as $eventname) {
+                    $data->$eventname = true;
+                }
             }
         }
         parent::set_data($data);